requirements.txt
# Core dependencies
# Note: Requires Python 3.10+

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
requests>=2.31.0
python-multipart>=0.0.6

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Statistics and analytics
scikit-learn>=1.3.0
statsmodels>=0.14.0
filterpy>=1.4.5

# Visualization
streamlit>=1.28.0
plotly>=5.18.0

# Database
# SQLite is built-in to Python

# Utilities
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

